Rename scenarios to model_definition and rearrange test location.
