+++ src/calliope/preprocess/__init__.py
@@ -2,5 +2,5 @@
 
 from calliope.preprocess.data_tables import DataTable
 from calliope.preprocess.model_data import ModelDataFactory
+from calliope.preprocess.model_definition import prepare_model_definition
 from calliope.preprocess.model_math import CalliopeMath
-from calliope.preprocess.scenarios import prepare_model_definition
