+++ tests/test_core_preprocess.py
@@ -5,7 +5,6 @@ import pytest
 
 import calliope
 import calliope.exceptions as exceptions
-from calliope.attrdict import AttrDict
 from calliope.io import read_rich_yaml
 
 from .common.util import build_test_model as build_model
@@ -17,8 +16,8 @@ class TestModelRun:
         """Test creating a model from dict/AttrDict instead of from YAML"""
         model_dir = data_source_dir.parent
         model_location = model_dir / "model.yaml"
-        model_dict = read_rich_yaml(model_location)
-        node_dict = AttrDict(
+        model_dict = calliope.io.read_rich_yaml(model_location)
+        node_dict = calliope.AttrDict(
             {
                 "nodes": {
                     "a": {"techs": {"test_supply_elec": {}, "test_demand_elec": {}}},
@@ -34,108 +33,6 @@ class TestModelRun:
 
         # test as dict
         calliope.Model(model_dict.as_dict())
-
-    @pytest.mark.filterwarnings(
-        "ignore:(?s).*(links, test_link_a_b_elec) | Deactivated:calliope.exceptions.ModelWarning"
-    )
-    def test_valid_scenarios(self, dummy_int):
-        """Test that valid scenario definition from overrides raises no error and results in applied scenario."""
-        override = read_rich_yaml(
-            f"""
-            scenarios:
-                scenario_1: ['one', 'two']
-
-            overrides:
-                one:
-                    techs.test_supply_gas.flow_cap_max: {dummy_int}
-                two:
-                    techs.test_supply_elec.flow_cap_max: {dummy_int/2}
-
-            nodes:
-                a:
-                    techs:
-                        test_supply_gas:
-                        test_supply_elec:
-                        test_demand_elec:
-            """
-        )
-        model = build_model(override_dict=override, scenario="scenario_1")
-
-        assert (
-            model._model_data.sel(techs="test_supply_gas")["flow_cap_max"] == dummy_int
-        )
-        assert (
-            model._model_data.sel(techs="test_supply_elec")["flow_cap_max"]
-            == dummy_int / 2
-        )
-
-    def test_valid_scenario_of_scenarios(self, dummy_int):
-        """Test that valid scenario definition which groups scenarios and overrides raises
-        no error and results in applied scenario.
-        """
-        override = read_rich_yaml(
-            f"""
-            scenarios:
-                scenario_1: ['one', 'two']
-                scenario_2: ['scenario_1', 'new_location']
-
-            overrides:
-                one:
-                    techs.test_supply_gas.flow_cap_max: {dummy_int}
-                two:
-                    techs.test_supply_elec.flow_cap_max: {dummy_int/2}
-                new_location:
-                    nodes.b.techs:
-                        test_supply_elec:
-
-            nodes:
-                a:
-                    techs:
-                        test_supply_gas:
-                        test_supply_elec:
-                        test_demand_elec:
-            """
-        )
-        model = build_model(override_dict=override, scenario="scenario_2")
-
-        assert (
-            model._model_data.sel(techs="test_supply_gas")["flow_cap_max"] == dummy_int
-        )
-        assert (
-            model._model_data.sel(techs="test_supply_elec")["flow_cap_max"]
-            == dummy_int / 2
-        )
-
-    def test_invalid_scenarios_dict(self):
-        """Test that invalid scenario definition raises appropriate error"""
-        override = read_rich_yaml(
-            """
-            scenarios:
-                scenario_1:
-                    techs.foo.bar: 1
-            """
-        )
-        with pytest.raises(exceptions.ModelError) as excinfo:
-            build_model(override_dict=override, scenario="scenario_1")
-
-        assert check_error_or_warning(
-            excinfo, "(scenarios, scenario_1) | Unrecognised override name: techs."
-        )
-
-    def test_invalid_scenarios_str(self):
-        """Test that invalid scenario definition raises appropriate error"""
-        override = read_rich_yaml(
-            """
-            scenarios:
-                scenario_1: 'foo'
-            """
-        )
-        with pytest.raises(exceptions.ModelError) as excinfo:
-            build_model(override_dict=override, scenario="scenario_1")
-
-        assert check_error_or_warning(
-            excinfo, "(scenarios, scenario_1) | Unrecognised override name: foo."
-        )
 
     def test_undefined_carriers(self):
         """Test that user has input either carrier or carrier_in/_out for each tech"""
